refactor: split dce tail sinking
